installed react-router-dom, adjusted App.jsx
installed react-router-dom.

changes to App.jsx:
now controls the flow using react router. previous functionality now follows the route /main into /create or /study. other implementations pending.

created /src/pages for an easier time routing. deleted some folders and components under /src/Components/Main, as they were becoming redundant.
