+++ client/package.json
@@ -9,6 +9,7 @@
     "lodash": "^4.17.21",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.15.0",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
