--- client/src/Components/Main/Main.module.css
@@ -1,7 +1,0 @@
-.main {
-  display: flex;
-  flex: 1;
-  position: relative;
-  flex-direction: column;
-  justify-content: center;
-}
