--- client/src/Components/Main/Modes/Create/Create.jsx
@@ -1,47 +1,0 @@
-import React, { useState } from "react";
-
-import ConfigContext from "../../../../store/config-context";
-import ProgressBar from "../../../UI/Progress/ProgressBar";
-import CreateSection from "./CreateSection";
-import CreateQuestions from "./CreateQuestions";
-import CreatePassages from "./CreatePassages";
-import CreateExtras from "./CreateExtras";
-import ConfirmRequestForm from "../../../UI/ConfirmRequest/ConfirmRequestForm";
-import Editor from "../../Editor/Editor";
-import classes from "./Create.module.css";
-
-const Create = (props) => {
-  const [configs, updateConfigs] = useState({
-    section: null,
-    questionTypes: null,
-    passageStyles: ["varied"],
-    passageTopics: ["varied"],
-    difficulty: ["varied"],
-    numberOfQuestions: 5,
-    wordsToUse: [],
-  });
-  const [activeConfig, setActiveConfig] = useState("section");
-
-  console.log(configs);
-
-  return (
-    <ConfigContext.Provider value={{ configs, updateConfigs, setActiveConfig }}>
-      {activeConfig !== "editor" && (
-        <ProgressBar
-          barFor="create"
-          active={activeConfig}
-          onSelect={setActiveConfig}
-          className={classes["progress-bar"]}
-        />
-      )}
-      {activeConfig === "section" && <CreateSection />}
-      {activeConfig === "questions" && <CreateQuestions />}
-      {activeConfig === "passages" && <CreatePassages />}
-      {activeConfig === "extras" && <CreateExtras />}
-      {activeConfig === "confirm" && <ConfirmRequestForm />}
-      {activeConfig === "editor" && <Editor />}
-    </ConfigContext.Provider>
-  );
-};
-
-export default Create;
