+++ client/src/index.css
@@ -16,9 +16,3 @@ body {
   /* display: flex;
   flex-flow: column; */
 }
-
-.main-container {
-  height: 100vh;
-  display: flex;
-  flex-direction: column;
-}
