+++ client/src/pages/ErrorPage/ErrorPage.module.css
@@ -1,0 +1,17 @@
+.error-page {
+  height: 100vh;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  gap: 48px;
+}
+
+.error-page h1 {
+  font-size: 9.6rem;
+  margin-bottom: 48px;
+}
+
+.error-page p {
+  font-size: 4.8rem;
+}
