+++ client/src/pages/MainInterface/MainInterface.module.css
@@ -1,0 +1,6 @@
+.main-container {
+  height: 100vh;
+  display: flex;
+  flex-direction: column;
+  min-height: 100vh;
+}
