+++ client/src/pages/ModeSelectionPage/Main.module.css
@@ -1,3 +1,11 @@
+.main {
+  display: flex;
+  flex: 1;
+  position: relative;
+  flex-direction: column;
+  justify-content: center;
+}
+
 .splash {
   display: flex;
   flex-direction: column;
